approvalsystem/requirements/requirement.txt
asgiref==3.4.1
Django==3.2.7
django-concurrency==2.3
django-cors-headers==3.8.0
django-environ==0.6.0
django-extensions==3.1.3
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
drf-yasg==1.20.0
